Move helpers to content
